tests/mediacapture-depth/MediaStream-interface-manual.https.html

<!DOCTYPE html>
<meta charset="utf-8">
<title>MediaStream Depth constructor algorithm</title>
<link rel="author" title="Intel" href="http://www.intel.com">
<link rel="help" href="https://w3c.github.io/mediacapture-main/#h-introduction">
<script src="../resources/testharness.js"></script>
<script src="../resources/testharnessreport.js"></script>
<meta name="flags" content="interact">

<h2>Preconditions</h2>
<ol>
  <li>
    Use a test device with depth camera(embedded or external) and an audio device.
  </li>
  <li>
    When prompted, accept to share your audio and depth streams.
  </li>
</ol>

<div id="log"></div>

<script>

async_test(t => {
  return navigator.mediaDevices.getUserMedia({video: {videoKind: {exact: "depth"}}, audio:true})
  .then(stream => {
    let stream1 = new MediaStream();
    assert_not_equals(stream.id, stream1.id, "Two different MediaStreams have different ids");

    let stream2 = new MediaStream(stream);
    assert_not_equals(stream.id, stream2.id, "A MediaStream constructed from another have different ids");

    let audioTrack = stream.getAudioTracks()[0];
    let videoTrack = stream.getVideoTracks()[0];
    assert_equals(audioTrack, stream2.getAudioTracks()[0], "A MediaStream constructed from another share the same audio track");
    assert_equals(videoTrack, stream2.getVideoTracks()[0], "A MediaStream constructed from another share the same video track");

    let stream3 = new MediaStream([videoTrack]);
    assert_equals(stream3.getTrackById(videoTrack.id), videoTrack, "a non-ended track gets added via the MediaStream constructor");

    var videoTrack_clone = videoTrack.clone();
    videoTrack_clone.addEventListener("ended", t.unreached_func("ended event should not be fired by MediaStreamTrack.stop()."));
    let value = videoTrack_clone.stop();
    assert_equals(value, undefined, "the stop function return a void type value");

    let stream4 = new MediaStream([videoTrack_clone, audioTrack]);
    assert_equals(stream4.getTrackById(videoTrack_clone.id), videoTrack_clone, "an ended track gets added via the MediaStream constructor");
    assert_equals(stream4.getTrackById(audioTrack.id), audioTrack, "a non-ended track gets added via the MediaStream constructor even if the previous track was ended");
    assert_true(stream4.active, "a MediaStream created using the MediaStream() constructor whose arguments are lists of MediaStreamTrack objects that are not all ended, the MediaStream object MUST be created with its active attribute set to true");

    var stream5 = new MediaStream([videoTrack_clone]);
    assert_equals(stream5.getTrackById(videoTrack_clone.id), videoTrack_clone, "an ended track gets added via the MediaStream constructor");
    assert_false(stream5.active, "a MediaStream created using the MediaStream() constructor whose arguments are lists of MediaStreamTrack objects that are all ended, the MediaStream object MUST be created with its active attribute set to false");

    // Use a timeout to detect a misfire of the ended event.
    t.step_timeout(t.step_func_done());
  });
}, "Tests that a MediaStream Depth constructor follows the algorithm set in the spec");

</script>
